+++ library/src/net/simonvt/widget/LeftDrawer.java
@@ -41,7 +41,7 @@ public class LeftDrawer extends MenuDrawer {
     protected void onLayout(boolean changed, int l, int t, int r, int b) {
         final int width = r - l;
         final int height = b - t;
-        final int offsetPixels = mOffsetPixels;
+        final int offsetPixels = (int) mOffsetPixels;
 
         mMenuContainer.layout(0, 0, mMenuSize, height);
         offsetMenu(offsetPixels);
@@ -157,12 +157,12 @@ public class LeftDrawer extends MenuDrawer {
 
     @Override
     protected void onMoveEvent(float dx) {
-        setOffsetPixels(Math.min(Math.max(mOffsetPixels + (int) dx, 0), mMenuSize));
+        setOffsetPixels(Math.min(Math.max(mOffsetPixels + dx, 0), mMenuSize));
     }
 
     @Override
     protected void onUpEvent(MotionEvent ev) {
-        final int offsetPixels = mOffsetPixels;
+        final int offsetPixels = (int) mOffsetPixels;
 
         if (mIsDragging) {
             mVelocityTracker.computeCurrentVelocity(1000, mMaxVelocity);
